requirements.txt
altgraph==0.17.4
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.2
pynput==1.8.1
pystray==0.19.5
pywin32==310
pywin32-ctypes==0.2.3
setuptools==78.1.0
six==1.17.0
ujson==5.10.0
